# planly-java/src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: Planly Contracts PH0 T1
  version: 0.1.0
  description: >-
    Stack-A-Monolith: single Worker + Static Assets. API served via Hono under
    /v1, web served from web/dist via /* fallback. B2-01: generate, plans,
    readjust, operations active with 202 {opId,status} + GET /operations/:opId
    + GET /plans/:id teaser-or-full + POST /plans/:id/readjust 201
    {readjustmentId,baseVersion,opId}. CUT (probe-gated): entitlement, Steps
    2-6, discard. Cookie draftId: 30d, Secure, HttpOnly,
    SameSite Lax, Path /. Allowlist: 9 ops, 7 active, 2 x-probe-gated.
    Methods allowlist: GET, POST, PUT, DELETE only. Naming: id, opId,
    Idempotency-Key, readjustmentId, version, n. Teaser allowlist frozen.
    Idempotency-Key UUIDv4 only else 400.
  x-stack: Stack-A-Monolith single Worker + Static Assets, Hono /v1 + web/dist /*
  x-scope:
    go: PH0 T1 drafts + steps/:n Step1 Q1-Q4 plus B2-01 generate/plans/readjust/operations
    cut:
      - entitlement
      - steps-2-to-6
      - discard
  x-cookie-draftId:
    name: draftId
    maxAgeDays: 30
    secure: true
    httpOnly: true
    sameSite: Lax
    path: /
  x-slos:
    - id: SLO-1
      inv: INV-1
      text: Draft create p95 < 500ms, availability 99.9%
    - id: SLO-2
      inv: INV-2
      text: Step1 read p95 < 300ms, availability 99.9%
    - id: SLO-3
      inv: INV-3
      text: Step1 write p95 < 500ms, durability 99.999999%
    - id: SLO-4
      inv: INV-3
...
    - id: SLO-5
      inv: INV-4
...
servers:
...
paths:
  /drafts:
    post:
      operationId: createDraft
      summary: Create draft (active)
      tags: [drafts]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDraftRequest'
      responses:
        '201':
          description: Draft created; sets draftId cookie
          headers:
            Set-Cookie:
              schema:
                type: string
                example: draftId=abc; Max-Age=2592000; Secure; HttpOnly; SameSite=Lax; Path=/
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Draft'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /drafts/{id}/generate:
    post:
      operationId: generateDraft
      summary: Generate plan from draft (async, idempotent)
      tags: [gated]
      parameters:
        - $ref: '#/components/parameters/DraftIdPath'
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateRequest'
      responses:
        '202':
          description: Generate accepted; poll GET /operations/{opId}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateAccepted'
        '200':
          description: Duplicate Idempotency-Key replay returns original opId
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateAccepted'
        '400':
          description: Missing or malformed Idempotency-Key (UUIDv4 only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '402':
          description: Entitlement required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Step stale or concurrent generate conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Step incomplete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /drafts/{id}:
    delete:
      operationId: discardDraft
      summary: Discard draft (CUT probe-gated)
      tags: [gated]
      x-probe-gated: true
      parameters:
        - $ref: '#/components/parameters/DraftIdPath'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '451':
          description: Probe-gated in PH0 T1
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  parameters:
    DraftIdPath:
      name: id
      in: path
      required: true
      description: Draft id (naming canonical)
      schema:
        $ref: '#/components/schemas/Id'
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Idempotency key, UUIDv4 only else 400
      schema:
        type: string
        format: uuid
        pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-4[0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$
        example: 550e8400-e29b-41d4-a716-446655440000
    IdempotencyKeyRequired:
      name: Idempotency-Key
      in: header
      required: true
      description: Idempotency key required on generate/readjust, UUIDv4 only else 400
      schema:
        type: string
        format: uuid
        pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-4[0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$
        example: 550e8400-e29b-41d4-a716-446655440000
...
  schemas:
    Id:
      type: string
      minLength: 1
      maxLength: 128
      example: dr_01HXYZ
    OpId:
      type: string
      description: Operation id canonical (opId)
      minLength: 1
      maxLength: 128
      example: op_01HXYZ
    Version:
      type: integer
      minimum: 1
      example: 1
...
    Draft:
      type: object
      required: [id, version]
      additionalProperties: false
      properties:
        id:
          $ref: '#/components/schemas/Id'
        version:
          $ref: '#/components/schemas/Version'
        opId:
          $ref: '#/components/schemas/OpId'
    CreateDraftRequest:
      type: object
      additionalProperties: false
      properties:
        opId:
          $ref: '#/components/schemas/OpId'
    GenerateRequest:
      type: object
      additionalProperties: false
      properties:
        opId:
          $ref: '#/components/schemas/OpId'
...
    GenerateAccepted:
      type: object
      required: [opId, status]
      additionalProperties: false
      properties:
        opId:
          $ref: '#/components/schemas/OpId'
        status:
          type: string
          const: accepted
...
    Error:
      type: object
      required: [code, message]
      additionalProperties: false
      properties:
        code:
          type: string
          example: gated
        message:
          type: string
          example: Probe-gated in PH0 T1
    # T1 wizard Steps 2-6 StepDef skeletons — static-string grade,
    # rendered-unverified. SSR Q1-Q4 (Step1) untouched. CUT LINE: no horizon
    # field (2-12mo is subscription-validity copy only), no sprint/daily
    # object shapes (loader/metric strings are display-only).
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...